% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha.globin}
\alias{alpha.globin}
\title{Deletion data from the alpha-globin locus}
\format{A data frame with 5 variables.
\describe{
  \item{condition}{A string describing the experimental condition}
  \item{expression}{the normalized expression level}
  \item{E*}{The 0/1 (absence/presence) coding for each enhancer}
}}
\source{
\url{http://www.nature.com/ng/journal/vaop/ncurrent/full/ng.3605.html}
}
\usage{
alpha.globin
}
\description{
The expression data from Hay et al. (2016) for their deletion of subsets of the alpha-globin enhancers.
Expression data is already normalized to the wild type. See paper for more details
}
\keyword{datasets}
