% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeActivity.R
\name{computeActivity}
\alias{computeActivity}
\title{Computes the value of the activity term}
\arguments{
\item{x}{enhancerDataObject}
}
\description{
Computes the value of the activity function for each row of the designMatrix
}
