% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParamPvals.R
\name{computeParamPvals}
\alias{computeParamPvals}
\title{Computes per parameter p-value}
\arguments{
\item{x}{enhancerDataObject}

\item{threads}{number of threads to use for refiting, requires parallel}

\item{...}{all aditional parameters to be passed to optimizer function \code{\link{optimDE}}}
}
\value{
A list with two elements:
\itemize{
  \item{pVals: A table of the coefficient p-values}
  \item{reducedModels: Refitted reduced models used to estimate the p-values}
}
}
\description{
Computes a p-value for the intercept and each enhancer coefficient in the link function by performing
likelihood ratio tests between the full model and a reduced model without that coefficient.
Can be parallelized if doParallel is installed. NOTE. Make sure your full model is already fit before
running this function!!!
}
\examples{
## Create a test design data.frame
design=expand.grid(E1=c(0,1),E2=c(0,1),E3=c(0,1))
## Create fake expression data
expression=c(0.001,0.2,0.3,0.7,0.4,0.8,0.6,1)*100
## Create activity function
actFun=formula(~E1+E2+E3+E1:E2)
edo = enhancerDataObject(expression,design,actFun)
edo=optimDE(edo,maxit=500)
param.stats=computeParamPvals(edo,maxit=500)
}
