% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerDataObject-class.R
\name{enhancerDataObject-class}
\alias{enhancerDataObject-class}
\title{Class enhancerDataObject}
\description{
Class \code{enhancerDataObject} holds knockout and expression data along with model parameters
}
