% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerDataObject-class.R
\name{enhancerDataObject}
\alias{enhancerDataObject}
\title{enhancerDataObject Constructor}
\usage{
enhancerDataObject(expressionData, designInfo, activityFunction,
  errorModel = c("lognormal", "gaussian"), linkFunction = c("additive",
  "exponential", "logistic"), activityParameterBounds = c(-10^3, 10^3),
  errorParameterBounds = c(10^-3, 10^3),
  scaleParameterBounds = c(10^-3, 10^3))
}
\arguments{
\item{expressionData}{A numeric vector with expression data, one entry for each row in the designInfo matrix}

\item{designInfo}{A 0/1 (absence/presence) coded matrix with one column for each enhancer and one row for each experiment}

\item{activityFunction}{the formula for the activity function (e.g. ~ E1 + E2 + E1*E2)}

\item{errorModel}{Desired error model for the regression function. Currently only lognormal or gaussian.}

\item{linkFunction}{Function that converts activity to gene expression}

\item{activityParameterBounds}{sets upper and lower bounds for coefficients of activity function - c(lower,upper)}

\item{errorParameterBounds}{sets upper and lower bounds for the error parameter - c(lower,upper)}

\item{scaleParameterBounds}{sets upper and lower bounds for the saturation parameter (only in logistic model) - c(lower,upper)}
}
\description{
Contructs a simple kinetic model object from synthesis and degredation rates
}
\examples{
## Create a test design data.frame
design=expand.grid(E1=c(0,1),E2=c(0,1),E3=c(0,1))
## Create fake expression data
expression=c(0.001,0.2,0.3,0.7,0.4,0.8,0.6,1)*100
## Create activity function
actFun=formula(~E1+E2+E3+E1:E2)
edo = enhancerDataObject(expression,design,actFun)

}
