% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorIntervals.R
\name{errorIntervals}
\alias{errorIntervals}
\title{Computes error bands for model}
\arguments{
\item{x}{enhancerDataObject}

\item{activity}{A range of activity values}

\item{quantiles}{The quantile bands to plot}
}
\description{
Computes error bands for model
}
