% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDE.R
\name{optimDE}
\alias{optimDE}
\title{Optimizes enhancer model with evolutionary algorithm}
\arguments{
\item{x}{enhancerDataObject}

\item{maxit}{the number of generations run}

\item{refine}{if TRUE, run gradient descent on the best solution from the evolutionary algorithm}
}
\description{
Runs differential evolution optimization method on enhancer model
}
\examples{
## Create a test design matrix
design = matrix(c(0,1,0,1,0,0,1,1),nrow=4)
colnames(design) = c("E1","E2")
design = as.data.frame(design)
## Create fake expression data
expression = c(0,0.2,0.3,0.9)
## Create activity function
actFun = formula(~E1+E2+E1:E2)
edo = enhancerDataObject(expression,design,actFun)
edo=optimDE(edo,refine=TRUE)

}
