% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictExpression.R
\name{predictExpression}
\alias{predictExpression}
\alias{predictExpression}
\title{Predicts expression}
\arguments{
\item{x}{enhancerDataObject}

\item{x}{enhancerDataObject}

\item{activity}{numeric vector of activity values}
}
\description{
Predicts expression from object activity values and link function

Predicts expression from supplied activity values and link function
}
\examples{


}
