% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wap}
\alias{wap}
\title{Deletion data from the wap locus}
\format{A data frame with 5 variables.
\describe{
  \item{condition}{A string describing the experimental condition}
  \item{expression}{the normalized expression level}
  \item{E*}{The 0/1 (absence/presence) coding for each enhancer}
}}
\source{
\url{http://www.nature.com/ng/journal/vaop/ncurrent/full/ng.3606.html}
}
\usage{
wap
}
\description{
The expression data from Shin et al. (2016) for their deletion of subsets of the Wap enhancers.
Expression data is already normalized to the wild type. See paper for more details
}
\keyword{datasets}
