% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeParamPvals.R
\name{computeParamPvals}
\alias{computeParamPvals}
\title{Computes per parameter p-value}
\arguments{
\item{x}{enhancerDataObject}

\item{threads}{number of threads to use for refiting, requires doParallel}

\item{...}{all aditional parameters to be passed to optimizer function \code{\link{optimDE}}}
}
\description{
Computes a p-value for the intercept and each enhancer coefficient in the link function by performing
likelihood ratio tests between the full model and a reduced model without that coefficient.
Can be parallelized if doParallel is installed. NOTE. Make sure your full model is already fit before
running this function!!!
}
\examples{
## Create a test design matrix
design = matrix(c(0,1,0,1,0,0,1,1),nrow=4)
colnames(design) = c("E1","E2")
design = as.data.frame(design)
## Create fake expression data
expression = c(0,0.2,0.3,0.9)
## Create activity function
actFun = formula(~E1+E2+E1*E2)
edo = enhancerDataObject(expression,design,actFun)
edo=optimDE(edo,maxit=300)
param.stats=computeParamPvals(edo,maxit=300)
}
