% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimDE.R
\name{optimDE}
\alias{optimDE}
\title{Optimizes enhancer model with evolutionary algorithm}
\arguments{
\item{x}{enhancerDataObject}

\item{maxit}{the number of generations run (overrides the value set by control)}

\item{refine}{if TRUE, run gradient descent on the best solution from the evolutionary algorithm}

\item{threads}{integer value for number of threads to use (overrides the value set by control)}

\item{control}{A list of control parameters that can be passed to \code{\link[DEoptim]{DEoptim}} (also see \code{\link[DEoptim]{DEoptim.control}})}
}
\description{
Runs differential evolution optimization method on enhancer model
}
\examples{
## Create a test design data.frame
design=expand.grid(E1=c(0,1),E2=c(0,1),E3=c(0,1))
## Create fake expression data
expression=c(0.001,0.2,0.3,0.7,0.4,0.8,0.6,1)*100
## Create activity function
actFun=formula(~E1+E2+E3+E1:E2)
edo = enhancerDataObject(expression,design,actFun)
edo=optimDE(edo,maxit=500,refine=TRUE)

}
