% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optimGD}
\alias{optimGD}
\title{Optimizes enhancer model}
\arguments{
\item{x}{enhancerDataObject}
}
\description{
Runs optimization method on enhancer model
}
\examples{
## Create a test design data.frame
design=expand.grid(E1=c(0,1),E2=c(0,1),E3=c(0,1))
## Create fake expression data
expression=c(0.001,0.2,0.3,0.7,0.4,0.8,0.6,1)*100
## Create activity function
actFun=formula(~E1+E2+E3+E1:E2)
edo = enhancerDataObject(expression,design,actFun)

}
