% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R, R/optimDE.R, R/plotModel.R
\name{optimMod}
\alias{optimMod}
\alias{optimMod}
\alias{optimMod}
\title{Optimizes enhancer model}
\arguments{
\item{x}{enhancerDataObject}

\item{x}{enhancerDataObject}

\item{x}{enhancerDataObject}
}
\description{
Runs optimization method on enhancer model

Runs differential evolution optimization method on enhancer model

Plots data and model prediction
}
\examples{



}
